Updating main file + .gitignore

+++ test.py
@@ -6,6 +6,7 @@ class TestProcessManager(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
+    def setUp(self):
         self.TEMP_FOLDER = 'temp-process'
         self.TEMP_BINARY = 'temp-dat-unittest'
         self.TEMP_PROCESS_NAME = 'testcase-test-1'
@@ -14,7 +15,7 @@ class TestProcessManager(unittest.TestCase):
         self.pm = server.ProcessManager(persistence_filename=self.TEMP_BINARY)
 
     def test_createProcess(self):
-        print('Creando proceso nuevo')
+        print('+\tCreando proceso nuevo')
         self.pm.createProcess(
             self.TEMP_PROCESS_NAME, 
             self.TEMP_PROCESS_FILENAME, 
@@ -25,22 +26,22 @@ class TestProcessManager(unittest.TestCase):
         self.assertEqual(self.TEMP_PROCESS_NAME, process.process_name)
     
     def test_getNotExistingProcess(self):
-        print('Intentar devolver un proceso que no existe')
+        print('+\tIntentar devolver un proceso que no existe')
         self.assertIsNone(self.pm.getProcessByName('testcase-process-90'))
        
     def test_runExistentProcess(self):
-        print('Ejecucion del proceso creado')
+        print('+\tEjecucion del proceso creado')
         self.pm.runProcess(self.TEMP_PROCESS_NAME)
         process = self.pm.getProcessByName(self.TEMP_PROCESS_NAME)
         time.sleep(1)
         self.assertIsNotNone(process.pid)
 
     def test_runNotExistentProcess(self):
-        print('Intentando ejecutar un proceso inexistente')
+        print('+\tIntentando ejecutar un proceso inexistente')
         self.assertFalse(self.pm.runProcess('not-exists-901924'))
 
     def test_stopRunningProcess(self):
-        print('Terminar ejecucion de proceso')
+        print('+\tTerminar ejecucion de proceso')
         process = self.pm.getProcessByName(self.TEMP_PROCESS_NAME)
         self.assertIsNotNone(process.pid)
         self.pm.stopProcess(self.TEMP_PROCESS_NAME)
@@ -49,12 +50,19 @@ class TestProcessManager(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    try:
-        os.remove(os.path.join('.','bin','temp-dat-unittest'))
+    if os.path.exists(os.path.join('.','temp-process', 'process-test-1.py')):
+        print('BORRANDO PROCESS')
         os.remove(os.path.join('.','temp-process', 'process-test-1.py'))
-    except:
-        pass
-    finally:
+    if os.path.exists(os.path.join('.','bin', 'temp-dat-unittest')):
+        print('BORRANDO DATA')
+        os.remove(os.path.join('.','bin', 'temp-dat-unittest'))
+    if os.path.isdir(os.path.join('.','temp-process')):
+        print('BORRANDO FILE')
+        os.rmdir(os.path.join('.','temp-process'))
+    
+    try:
         os.mkdir('temp-process')
+    except Exception as err:
+        print('Error al crear directorio:', err)
+    finally:
         unittest.main()
-        shutil.rmtree('temp-process')
